Draw locally verified checkmark over Sourcify logo
