+++ src/execution/components/DecoratedAddressLink.tsx
@@ -1,3 +1,4 @@
+import { Metadata } from "@ethereum-sourcify/lib-sourcify";
 import {
   faBomb,
   faBurn,
@@ -6,12 +7,13 @@ import {
   faStar,
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { FC, memo, useContext } from "react";
+import { FC, memo, useContext, useEffect, useState } from "react";
 import { NavLink } from "react-router";
 import { resolverRendererRegistry } from "../../api/address-resolver";
 import AddressLegend from "../../components/AddressLegend";
 import SourcifyLogo from "../../sourcify/SourcifyLogo";
 import { useSourcifyMetadata } from "../../sourcify/useSourcify";
+import CheckedContractStorage from "../../storage/CheckedContractStorage";
 import { AddressContext, ChecksummedAddress, ZERO_ADDRESS } from "../../types";
 import { useResolvedAddress } from "../../useResolvedAddresses";
 import { RuntimeContext } from "../../useRuntime";
@@ -47,7 +49,26 @@ const DecoratedAddressLink: FC<DecoratedAddressLinkProps> = ({
   plain,
 }) => {
   const { config, provider } = useContext(RuntimeContext);
-  const match = useSourcifyMetadata(address, provider._network.chainId);
+  const chainId = provider._network.chainId;
+  const match = useSourcifyMetadata(address, chainId);
+  const [locallyVerified, setLocallyVerified] = useState<boolean>(false);
+
+  useEffect(() => {
+    // Check whether we locally verified this contract
+    if (match) {
+      const savedMetadataHash = CheckedContractStorage.get(chainId, address);
+      if (savedMetadataHash !== null) {
+        // TODO: Find reason our Metadata type is not compatible with Sourcify's
+        CheckedContractStorage.hashMetadata(
+          match.metadata as unknown as Metadata,
+        ).then((metadataHash) => {
+          if (metadataHash === savedMetadataHash) {
+            setLocallyVerified(true);
+          }
+        });
+      }
+    }
+  }, [match]);
 
   const mint = addressCtx === AddressContext.FROM && address === ZERO_ADDRESS;
   const burn = addressCtx === AddressContext.TO && address === ZERO_ADDRESS;
@@ -98,7 +119,7 @@ const DecoratedAddressLink: FC<DecoratedAddressLinkProps> = ({
           className="flex shrink-0 items-center self-center"
           to={`/address/${address}/contract`}
         >
-          <SourcifyLogo />
+          <SourcifyLogo locallyVerified={locallyVerified} />
         </NavLink>
       )}
       {plain ? (
