+++ src/sourcify/SourcifyLogo.tsx
@@ -1,14 +1,38 @@
+import { faCheckCircle, faCircle } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
 import SourcifyIcon from "./sourcify.svg";
 
-const SourcifyLogo: React.FC = () => (
-  <img
-    src={SourcifyIcon}
-    alt="Sourcify logo"
-    title="Verified by Sourcify"
-    width={16}
-    height={16}
-  />
-);
+interface SourcifyLogoProps {
+  locallyVerified?: boolean;
+}
+
+const SourcifyLogo: React.FC<SourcifyLogoProps> = ({ locallyVerified }) => {
+  const title = locallyVerified ? "Locally verified" : "Verified by Sourcify";
+  return (
+    <div className="relative inline-block">
+      <img
+        src={SourcifyIcon}
+        alt="Sourcify logo"
+        title={title}
+        width={16}
+        height={16}
+      />
+      {locallyVerified && (
+        <div title={title}>
+          {/* Draws a white circle behind to ensure the checkmark is white */}
+          <FontAwesomeIcon
+            className="absolute bottom-0 right-[-2px] ml-1 text-white text-[7pt]"
+            icon={faCircle}
+          />
+          <FontAwesomeIcon
+            className="absolute bottom-0 right-[-2px] ml-1 text-emerald-500 text-[8pt]"
+            icon={faCheckCircle}
+          />
+        </div>
+      )}
+    </div>
+  );
+};
 
 export default SourcifyLogo;
